@@ -33,7 +33,8 @@
     "handlebars": "^4.0.6",
     "hbs": "^4.0.1",
     "isomorphic-fetch": "^2.2.1",
-    "react": "^15.4.2",
+    "prop-types": "^15.5.4",
+    "react": "^15.5.3",
     "react-dom": "^15.4.2",
     "react-redux": "^5.0.3",
     "react-router": "^4.0.0",
